Build queue manipulation
Allow manipulation of the positions building sites are in the queue to be built.

Although this works, it has been done incorrectly, and needs to use proper game commands to keep with the code design  of the project.

(This was one of the first thins I attempted to change when I first played,  and had no knowledge of the source code)