Road network optimizations
This forces a reroute of roads that are unable to set a flag on it due to another being close by.

It will periodically check the building site until it is satisfied with the route, then will toggle a flag to stop being checked.

Due to roads being cut, the area is updated to check for breaks in the road network, and will attempt to fix them.