Main AI upgrade
This is the result of many many hours of testing.

All previous commits lead to this one.

This essentially sets a build script for the AI with a few restrictions.

It also copes with low starting resources.

The general idea is to ensure the AI is able to produce tools ASAP and a solid wood production is setup.

It also checks a certain amount of expansion has happened, so it doesn't try and build 50 buildings inside a settlement with 3 military buildings occupied.

Checks are also made if the resources are en-route to the building sites, allowing progression of the script earlier, rather than waiting for buildings to be fully built.

Once the first part of the script is done, we base most of the buildings we want on the size of our territory, OR the amount of spare Generals we have available.  This has the effect of slowing the AI's rapid expansion, but ensures Generals are being produced.