Placement of foresters improved.
When placing a forester, if we are in the early stages of the game,  the AI will attempt to place a forester next to a random woodcutter building or site.

In the later stages, the AI will look for heavy deforestation, due to too many woodcutters in the area, and place another forester near by.