Send builders to building sites depending on the priority set in build orders
Build orders is a great addition, but it's not without one major issue.

Builders will work at buildings depending on the order the building was connected in.

What we end up with is, a load of building sites with resources, but no builder present,  and on the other spectrum, builders present but no resources.

This code "should" help mitigate this scenario, yes it's ugly, but it's more a proof of concept.