Look for occupied buildings
Ensuring a good supply of building materials was a good start, but issues started to arise with tools.

The AI got to a point that it would cripple itself simply because too many mines were built and not enough miners were made.

This update will slow the build progression by making sure miners are actually present in at least 2 iron and 2 coal mines.

It then starts to build a basic food industry and also halts progression until enough have workers present.

Slowing down building and checking mines, ensures the toolsmiths have a reliable source of materials, which allow it to keep up with building.